+++ packages/use/index.ts
@@ -5,3 +5,4 @@ export * from './src/useCountdown/index';
 export * from './src/deepUnref/index';
 export * from './src/useRequest/index';
 export * from './src/createNamedSharedComposable/index';
+export * from './src/useDraggableDistance/index';